data_loader_service/requirements.txt
beautifulsoup4==4.13.3
camelot_py==1.0.0
numpy==2.2.3
pandas==2.2.3
pdfplumber==0.11.5
python-dotenv==1.0.1
Requests==2.32.3
